+++ aoc2024/day17/Registers.cs
@@ -2,8 +2,8 @@ namespace Advent_of_Code_2024.day17;
 
 public record Registers
 {
-    public int A { get; set; }
-    public int B { get; set; }
-    public int C { get; set; }
+    public long A { get; set; }
+    public long B { get; set; }
+    public long C { get; set; }
     public int InstructionPointer { get; set; }
 }
